Moved SSDP discovery code into minissdpc.c
